requirements.txt
flask
simple-pid
jsonpickle
RPi.GPIO
adafruit-circuitpython-max31865
adafruit-blinka
